requirements.txt
alembic==1.7.5
Flask==2.0.2
Flask_Migrate==3.1.0
Flask_RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
marshmallow==3.14.1
marshmallow_sqlalchemy==0.27.0
pytest==6.2.5
python-dotenv==0.19.2
setuptools==58.5.3
mysql-connector-python==8.0.27
coverage==6.2
coveralls==3.3.1
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6